pyproject.toml
[tool.poetry]
name = "discord-election-bot"
version = "0.1.0"
description = ""
authors = ["hemanta212 <sharmahemanta.212@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
python-dotenv = "^0.20.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
ptpython = "^3.0.20"
rich = "^12.4.1"
black = "^22.3.0"
loguru = "^0.6.0"
rope = "^1.0.0"
pylint = "^2.13.9"
python-lsp-server = "^1.4.1"
pylsp-mypy = "^0.5.7"
pyls-isort = "^0.2.2"
pylsp-rope = "^0.1.9"
ipython = "^8.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
